removed old get_auction_stats script
