+++ trufflescripts/get_auction_stats.js
@@ -1,7 +1,5 @@
 /* eslint no-console:0 */
-const DutchExchange = artifacts.require('DutchExchange')
-const TokenETH = artifacts.require('EtherToken')
-const TokenGNO = artifacts.require('TokenGNO')
+const { getTokenBalances, getTokenDeposits, deployed, getAllStatsForTokenPair } = require('./utils/contracts')(artifacts)
 const { getTime } = require('./utils')(web3)
 
 const getTimeStr = (timestamp) => {
@@ -13,9 +11,11 @@ const getTimeStr = (timestamp) => {
   return `${hh ? `${hh} hour(s) ` : ''}${mm ? `${mm} minute(s) ` : ''}${ss ? `${ss} second(s) ` : ''}`
 }
 
+const getNumDenStr = ([num, den]) => `${num}/${den} = ${(num / den).toFixed(8)}`
+
 /**
  * truffle exec trufflescripts/get_auction_stats.js
- * prints stats for the current and past auctions
+ * prints stats for the current and past ETH -> GNO auctions
  */
 
 /* eslint no-console: 0 */
@@ -26,90 +26,138 @@ module.exports = async () => {
     TESTS WILL NOT WORK IF PRICE_ORACLE DOES NOT YET SET A USD VALUE FOR ETHER!
     --------------------------------------------------------------------------
   `)
-  const dx = await DutchExchange.deployed()
-  const eth = await TokenETH.deployed()
-  const gno = await TokenGNO.deployed()
+  const { dx, eth, gno } = await deployed
 
-  const dxETHBalance = (await eth.balanceOf.call(dx.address)).toNumber()
-  const dxGNOBalance = (await gno.balanceOf.call(dx.address)).toNumber()
+  const { ETH: dxETH, GNO: dxGNO, TUL: dxTUL, OWL: dxOWL } = await getTokenBalances(dx.address)
 
-  console.log(`Auction holds:\t${dxETHBalance} ETH\t${dxGNOBalance} GNO`)
+  console.log(`Exchange holds:\t${dxETH} ETH\t${dxGNO} GNO\t${dxTUL} TUL\t${dxOWL} OWL`)
 
-  const auctionStart = (await dx.auctionStarts.call(eth.address, gno.address)).toNumber()
-  const now = getTime()
-
-  const timeUntilStart = auctionStart - now
-  const timeStr = getTimeStr(timeUntilStart * 1000)
-  // console.log(timeStr)
-  const auctionIndex = (await dx.latestAuctionIndices.call(eth.address, gno.address)).toNumber()
-  // console.log(auctionIndex)
   const [, seller, buyer] = web3.eth.accounts
 
-  const sellVolumeCurrent = (await dx.sellVolumes.call(eth.address, gno.address, auctionIndex)).toNumber()
-  const sellVolumeNext = (await dx.sellVolumes.call(eth.address, gno.address, auctionIndex + 1)).toNumber()
-  const sellerBalanceNext = sellVolumeNext &&
-    (await dx.sellerBalances.call(eth.address, gno.address, auctionIndex + 1, seller)).toNumber()
+  const [sellerDeposits, buyerDeposits] = await Promise.all([
+    getTokenDeposits(seller),
+    getTokenDeposits(buyer),
+  ])
 
-  console.log(`
-    Current auction index ${auctionIndex}
-    ______________________________________
-    now:\t\t\t${new Date(now * 1000).toTimeString()}
-    auctionStart:\t\t${new Date(auctionStart * 1000).toTimeString()}
-    ${timeUntilStart > 0 ? `starts in\t\t${timeStr}` : timeUntilStart < 0 ? `started\t\t${timeStr}ago` : 'just started'}
-    
-    sellVolumeCurrent:\t${sellVolumeCurrent}
-    sellVolumeNext:\t${sellVolumeNext}${sellerBalanceNext ? `\n  sellerBalance for next auction:\t${sellerBalanceNext}` : ''}  
-  `)
+  console.log('Deposits in the Exchange')
+  console.log(`  Seller:\t${sellerDeposits.ETH}\tETH,\t${sellerDeposits.GNO}\tGNO`)
+  console.log(`  Buyer:\t${buyerDeposits.ETH}\tETH,\t${buyerDeposits.GNO}\tGNO,`)
 
+  const now = getTime()
+  const stats = await getAllStatsForTokenPair({ sellToken: eth, buyToken: gno, accounts: [seller, buyer] })
 
-  // if auctionIndex === 3, indexes = [3, 2, 1]
-  const indexes = Array.from({ length: auctionIndex }, (v, i) => auctionIndex - i)
+  const {
+    // TODO: remove = [1, 1] workaround for ETH token when dx.priceOracle() changes
+    sellTokenOraclePrice = [1, 1],
+    buyTokenOraclePrice,
+    latestAuctionIndex,
+    auctionStart,
+    arbTokens,
+    sellVolumeCurrent,
+    sellVolumeNext,
+    buyVolume,
+    auctions,
+  } = stats
 
-  const readStats = async (i) => {
-    const buyVolume = (await dx.buyVolumes.call(eth.address, gno.address, i)).toNumber()
+  console.log('\nAuction pair ETH -> GNO')
 
-    let price, amountToClearAuction, timeUntilAuctionClears
-    try {
-      const [num, den] = (await dx.getPrice.call(eth.address, gno.address, i)).map(n => n.toNumber())
-      console.log(`DENOMINATOR = ${den}, NUMERATOR = ${num}`)
-      price = `1 ETH = ${(num / den).toFixed(8)} GNO`
-
-      // if current running auction
-      if (i === auctionIndex) {
-        /* eslint-disable no-mixed-operators */
-        amountToClearAuction = Math.floor(sellVolumeCurrent * num / den) - buyVolume
-        const timeWhenAuctionClears = Math.ceil(72000 * sellVolumeCurrent / buyVolume - 18000 + auctionStart)
-
-        timeUntilAuctionClears = getTimeStr((now - timeWhenAuctionClears) * 1000)
-      }
-    } catch (error) {
-      price = 'unavailable, auction hasn\'t started'
-
-      const [num, den] = (await dx.getPrice.call(eth.address, gno.address, i - 1)).map(n => n.toNumber())
-      price += `\n  last closingPrice:\t1 ETH = ${(num / den).toFixed(8)} GNO`
-    }
-
-    const sellerBalance = (await dx.sellerBalances.call(eth.address, gno.address, i, seller)).toNumber()
-    const buyerBalance = (await dx.buyerBalances.call(eth.address, gno.address, i, buyer)).toNumber()
-    // const sellerClaimed = (await dx.claimedAmounts(i, seller)).toNumber()
-    // const buyerClaimed = (await dx.claimedAmounts(i, buyer)).toNumber()
-
-    if (i !== auctionIndex) {
-      console.log('=============================')
-      console.log(`Auction index ${i}`)
-    }
-
-    console.log(`
-      buyVolume:\t\t${buyVolume}
-      price:\t\t${price}${amountToClearAuction ? `\n  to clear auction buy\t${amountToClearAuction} GNO` : ''}
-      ${timeUntilAuctionClears ? `will clear with time in ${timeUntilAuctionClears}` : ''}
-
-      sellerBalance:  ${sellerBalance}\tclaimed:  ${'SellerClaimed'} ETH
-      buyerBalance:   ${buyerBalance}\tclaimed:  ${'buyerClaimed'} GNO
+  if (sellTokenOraclePrice && buyTokenOraclePrice) {
+    console.log(`Oracle prices:
+    1 ETH = ${getNumDenStr(sellTokenOraclePrice)} ETH
+    1 GNO = ${getNumDenStr(buyTokenOraclePrice)} ETH
     `)
   }
 
-  for (const i of indexes) {
-    await readStats(i)
+  console.log(`
+    sellVolumeCurrent:\t${sellVolumeCurrent}
+    sellVolumeNext:\t${sellVolumeNext}
+    buyVolume:\t${buyVolume}
+  `)
+
+  console.log(`Arbitrage tokens:\t${arbTokens}`)
+  console.log(`latestAuctionIndex:\t${latestAuctionIndex}`)
+
+  console.log(`now:\t\t\t${new Date(now * 1000).toTimeString()}`)
+
+  if (auctionStart === 0) {
+    console.log('auction has never run before')
+  } else {
+    const timeUntilStart = auctionStart - now
+    const timeStr = getTimeStr(timeUntilStart * 1000)
+
+    if (timeUntilStart > 0) {
+      console.log(`next auction starts in\t\t${timeStr}`)
+    } else if (timeUntilStart < 0) {
+      console.log(`auction started\t\t${timeStr}ago`)
+    } else {
+      console.log('auction just started')
+    }
+  }
+
+  console.log('======================================')
+
+  for (const auctionStats of auctions) {
+    const {
+      auctionIndex,
+      closingPrice,
+      // sellVolume,
+      // buyVolume,
+      extraTokens,
+      isLatestAuction,
+      accounts,
+      price,
+    } = auctionStats
+
+    console.log(`
+    Auction index ${auctionIndex} ${isLatestAuction ? '(latest)' : ''}
+    ______________________________________
+    
+    extraTokens:\t${extraTokens}
+  `)
+
+    let closingPriceStr
+
+    if (closingPrice.some(n => n > 0)) {
+      closingPriceStr = `1 ETH = ${getNumDenStr(closingPrice)} GNO`
+    } else {
+      closingPriceStr = 'N/A'
+    }
+
+    console.log(`    closingPrice: ${closingPriceStr}`)
+
+    if (price) console.log(`\n  currentPrice: 1 ETH = ${getNumDenStr(price)} GNO`)
+
+    if (isLatestAuction && price && sellTokenOraclePrice && buyTokenOraclePrice) {
+      const [num, den] = price
+      const [sellTokenNum] = sellTokenOraclePrice
+      const [, buyTokenDen] = buyTokenOraclePrice
+
+      const amountToClearAuction = Math.floor((sellVolumeCurrent * num) / den) - buyVolume
+
+      if (amountToClearAuction > 0) console.log(`  to clear auction buy\t${amountToClearAuction} GNO`)
+
+      // const timeWhenAuctionClears = Math.ceil(72000 * sellVolumeCurrent / buyVolume - 18000 + auctionStart)
+      const timeWhenAuctionClears = Math.ceil((86400 / sellTokenNum / buyTokenDen) + auctionStart)
+      const timeUntilAuctionClears = getTimeStr((now - timeWhenAuctionClears) * 1000)
+
+      if (now - timeWhenAuctionClears >= 0) {
+        console.log(`  will clear with time in ${timeUntilAuctionClears}`)
+      } else {
+        console.log(`  cleared ${timeWhenAuctionClears} ago`)
+      }
+    }
+
+    if (accounts && Object.keys(accounts).length) {
+      console.log('\n\tsellerBalance,\tbuyerBalance,\tclaimedAmount')
+      for (const account of Object.keys(accounts)) {
+        const { sellerBalance, buyerBalance, claimedAmount } = accounts[account]
+
+        const accountName = account === seller ? 'Seller' : account === buyer ? 'Buyer' : account
+
+        console.log(`  ${accountName}:\t${sellerBalance},\t\t${buyerBalance},\t\t${claimedAmount}`)
+      }
+    }
+
+    console.log('=============================')
   }
 }
